Create justfile with initial k8s cluster setup with kind
Run `just create-cluster` to create the cluster and deploy flux
components into the created cluster. flux will be set up to watch this
repo by default.

Run `just delete-cluster` to delete the whole thing again. Need to
figure out what to do with persistent storage in the future though as
that needs to outlive the cluster.